move reader drawer stuff into drawer dir
