+++ frontend/src/components/reader/drawer/download-panel.vue
