+++ frontend/src/components/reader/drawer/keyboard-shortcuts-panel.vue
