+++ frontend/src/components/reader/drawer/reader-settings-panel.vue
